split plugin out of FluxNoteEditor component like Dylan had done
